+++ server/locale/FR/validOptions.js
@@ -11,12 +11,10 @@ export default {
             size: [Types.NUMBER, [12, 10, 11, 13, 14, 15, 16]],
             align: [Types.STRING, ['left', 'right', 'center']],
             fontFamily: [Types.STRING]
-        },
-        preset: [Types.STRING, [undefined, 'smallest']]
+        }
     },
     flex: {
         color: [Types.STRING, ['blue', 'black', 'white', 'gray|grey', 'monochrome', 'grayscale|greyscale']],
-        ratio: [Types.STRING, ['1x1', '1x4', '8x1', '20x1']],
-        preset: [Types.STRING, [undefined, 'smallest']]
+        ratio: [Types.STRING, ['1x1', '1x4', '8x1', '20x1']]
     }
 };