+++ src/components/modal/lib/hooks/calculator.js
@@ -1,7 +1,7 @@
-import { useReducer, useEffect, useMemo, useRef } from 'preact/hooks';
+import { useReducer, useMemo, useRef } from 'preact/hooks';
 import { debounce } from 'belter/src';
 
-import { useXProps, useServerData } from '../../../lib';
+import { useXProps, useServerData, useDidUpdateEffect } from '../../../lib';
 import { getContent } from '../utils';
 
 const reducer = (state, action) => {
@@ -60,7 +60,17 @@ const localize = (country, amount) => {
 export default function useCalculator({ autoSubmit = false } = {}) {
     const calculateRef = useRef();
     const { terms: initialTerms, country, setServerData } = useServerData();
-    const { currency, payerId, clientId, merchantId, onCalculate, amount, buyerCountry } = useXProps();
+    const {
+        currency,
+        payerId,
+        clientId,
+        merchantId,
+        onCalculate,
+        buyerCountry,
+        ignoreCache,
+        amount,
+        stageTag
+    } = useXProps();
     const [state, dispatch] = useReducer(reducer, {
         inputValue: localize(country, initialTerms.amount),
         prevValue: localize(country, initialTerms.amount),
@@ -77,7 +87,9 @@ export default function useCalculator({ autoSubmit = false } = {}) {
             payerId,
             clientId,
             merchantId,
-            buyerCountry
+            buyerCountry,
+            ignoreCache,
+            stageTag
         })
             .then(data => {
                 setServerData(data);
@@ -101,12 +113,22 @@ export default function useCalculator({ autoSubmit = false } = {}) {
             });
     };
 
-    // Automatically fetch terms when props change
-    useEffect(() => {
-        if (localize(country, amount) !== state.inputValue) {
-            fetchTerms(amount);
+    // Update the terms in the reducer based on outside changes to serverData
+    useDidUpdateEffect(() => {
+        // When amount xprop changes, Container.jsx will fetch new serverData (including terms)
+        // If we see new terms, which match the amount prop, but the value in the input does not match
+        // This means the amount changed outside the modal, so we update the terms
+        // we want to update the inputValue, so force autoSubmit: false
+        if (Number(initialTerms.amount) === amount && delocalize(country, state.inputValue) !== amount) {
+            dispatch({
+                type: 'terms',
+                data: {
+                    ...initialTerms,
+                    autoSubmit: false
+                }
+            });
         }
-    }, [payerId, clientId, merchantId, country, amount]);
+    }, [initialTerms, amount]);
 
     // Because we use state in this function, which changes every dispatch,
     // and we want it debounced, we need to use a ref to hold the most up-to-date function reference
@@ -115,7 +137,7 @@ export default function useCalculator({ autoSubmit = false } = {}) {
 
         if (state.prevValue !== state.inputValue && delocalizedValue !== 'NaN') {
             onCalculate({ value: delocalizedValue });
-            fetchTerms(delocalizedValue);
+            fetchTerms(delocalizedValue, autoSubmit);
         } else {
             // The input value may have changed, but the actual amount value did not
             // ex: $10.9 === $10.90