+++ tests/functional/spec/utils/modalSnapshot.js
@@ -23,6 +23,7 @@ expect.extend({ toMatchLargeSnapshot, toMatchSmallSnapshot });
 const modalSnapshot = async (testNameParts, viewport, account) => {
     const elementModal = await page.$(selectors.modal.iframe);
     const modalFrame = await elementModal.contentFrame();
+    await modalFrame.waitForSelector(selectors.modal.wrapper, { visible: true });
     const modalDimensions = await modalFrame.$eval(selectors.modal.contentBackground, element => ({
         x: element.offsetLeft,
         y: element.offsetTop,