+++ tests/unit/spec/src/components/message/Message.test.js
@@ -1,13 +1,11 @@
-/** @jsx h */
-import { h } from 'preact';
-import { render, fireEvent, waitFor, act } from '@testing-library/preact';
+import { getByText, fireEvent, queryByText } from '@testing-library/dom';
 
 import Message from 'src/components/message/Message';
-import { request, getOrCreateStorageID } from 'src/utils';
+import { request, getOrCreateStorageID, createState } from 'src/utils';
 import xPropsMock from 'utils/xPropsMock';
-import zoidComponentWrapper from 'utils/zoidComponentWrapper';
 
 jest.mock('src/utils', () => ({
+    createState: jest.fn(obj => [obj, jest.fn()]),
     getActiveTags: jest.fn(),
     getOrCreateStorageID: jest.fn(() => 'uid_26a2522628_mtc6mjk6nti'),
     request: jest.fn(() =>
@@ -29,7 +27,7 @@ jest.mock('src/utils', () => ({
     ppDebug: jest.fn(() => console.log('PayPal Debug Message'))
 }));
 
-describe('<Message />', () => {
+describe('Message', () => {
     const updateProps = xPropsMock({
         amount: 100,
         currency: 'USD',
@@ -44,14 +42,14 @@ describe('<Message />', () => {
         resize: jest.fn()
     });
 
-    const wrapper = zoidComponentWrapper({
+    const serverData = {
         markup: '<div>test</div>',
         meta: {
             messageRequestId: '12345'
         },
         parentStyles: 'body { color: black; }',
         warnings: []
-    });
+    };
 
     afterEach(() => {
         window.xprops.onClick.mockClear();
@@ -59,18 +57,34 @@ describe('<Message />', () => {
         window.xprops.onClick.mockClear();
         window.xprops.onMarkup.mockClear();
 
+        createState.mockClear();
         request.mockClear();
         getOrCreateStorageID.mockClear();
+        xPropsMock.clear();
+    });
+
+    test('Renders the button with styles', () => {
+        const button = Message(serverData);
+        const styles = [];
+        Object.keys(button.style).forEach(row => {
+            if (!Number.isNaN(Number(row)) && button.style[row] !== undefined) {
+                styles.push(button.style[row]);
+            }
+        });
+
+        expect(button instanceof HTMLButtonElement).toBe(true);
+        expect(JSON.stringify(styles)).toBe(
+            JSON.stringify(['display', 'background', 'padding', 'outline', 'text-align', 'font-family'])
+        );
     });
 
     test('Renders the server markup', () => {
-        const { getByText } = render(<Message />, { wrapper });
-
-        expect(getByText(/test/i)).toBeInTheDocument();
+        const messageDocument = document.body.appendChild(Message(serverData));
+        expect(getByText(messageDocument, /test/i)).toBeInTheDocument();
     });
 
     test('Fires onReady xProp after render', () => {
-        render(<Message />, { wrapper });
+        Message(serverData);
 
         expect(window.xprops.onReady).toHaveBeenCalledTimes(1);
         expect(window.xprops.onReady).toHaveBeenLastCalledWith({
@@ -82,11 +96,9 @@ describe('<Message />', () => {
     });
 
     test('Fires onClick xProp when clicked', () => {
-        const { container } = render(<Message />, { wrapper });
-        const button = container.firstChild;
+        const button = Message(serverData);
 
         fireEvent.click(button);
-
         expect(window.xprops.onClick).toHaveBeenCalledTimes(1);
         expect(window.xprops.onClick).toHaveBeenLastCalledWith({
             meta: {
@@ -96,8 +108,7 @@ describe('<Message />', () => {
     });
 
     test('Fires onHover xProp when hovered', () => {
-        const { container } = render(<Message />, { wrapper });
-        const button = container.firstChild;
+        const button = Message(serverData);
 
         fireEvent.mouseOver(button);
 
@@ -110,17 +121,19 @@ describe('<Message />', () => {
     });
 
     test('Fires onMarkup and onReady on complete re-render', async () => {
-        const { getByText, queryByText } = render(<Message />, { wrapper });
+        const messageDocument = document.body.appendChild(Message(serverData));
 
         expect(request).not.toHaveBeenCalled();
-        expect(getByText(/test/i)).toBeInTheDocument();
+        expect(getByText(messageDocument, /test/i)).toBeInTheDocument();
         expect(window.xprops.onReady).toHaveBeenCalledTimes(1);
+
         expect(window.xprops.onReady).toHaveBeenLastCalledWith({
             meta: {
                 messageRequestId: '12345'
             },
             deviceID: 'uid_26a2522628_mtc6mjk6nti'
         });
+
         expect(window.xprops.onMarkup).toHaveBeenCalledTimes(1);
         expect(window.xprops.onMarkup).toHaveBeenLastCalledWith({
             meta: {
@@ -130,18 +143,17 @@ describe('<Message />', () => {
             warnings: []
         });
 
-        act(() => {
-            updateProps({ amount: 200 });
-        });
+        updateProps({ amount: 200 });
 
-        await waitFor(() => {
-            expect(window.xprops.onReady).toHaveBeenCalledTimes(2);
-            expect(window.xprops.onMarkup).toHaveBeenCalledTimes(2);
-        });
+        // has to wait for the button to re-render
+        await new Promise(resolve => setTimeout(resolve, 0));
 
-        expect(queryByText(/test/i)).toBeNull();
-        expect(getByText(/mock/i)).toBeInTheDocument();
+        expect(window.xprops.onReady).toHaveBeenCalledTimes(2);
+        expect(window.xprops.onMarkup).toHaveBeenCalledTimes(2);
+        expect(queryByText(messageDocument, /test/i)).toBeNull();
+        expect(getByText(messageDocument, /mock/i)).toBeInTheDocument();
         expect(request).toHaveBeenCalledTimes(1);
+
         expect(window.xprops.onReady).toHaveBeenLastCalledWith({
             meta: {
                 messageRequestId: '23456'
@@ -160,7 +172,7 @@ describe('<Message />', () => {
     test('Passed deviceID from iframe storage to callback', () => {
         getOrCreateStorageID.mockReturnValue('uid_1111111111_11111111111');
 
-        render(<Message />, { wrapper });
+        Message(serverData);
 
         expect(window.xprops.onReady).toBeCalledWith({
             meta: {