+++ tests/unit/spec/src/components/modal/lib/providers/transition.test.js
@@ -17,11 +17,7 @@ jest.mock('src/components/lib/providers/xprops');
 describe('transition', () => {
     useXProps.mockReturnValue(defaultXProps);
 
-    afterEach(() => {
-        defaultXProps.onProps.mockClear();
-    });
-
-    test('useTransitionState updates when modal opens and can trigger closing', async () => {
+    test('useTransitionState can trigger modal close', async () => {
         const wrapper = ({ children }) => <TransitionStateProvider>{children}</TransitionStateProvider>;
         const { result } = renderHook(
             () => {
@@ -32,43 +28,16 @@ describe('transition', () => {
             { wrapper }
         );
 
-        expect(result.current.status).toBe(STATUS.CLOSED);
-
-        const [onProps] = defaultXProps.onProps.mock.calls[0];
-
-        onProps({ visible: true });
-
-        await new Promise(resolve => {
-            setTimeout(() => {
-                expect(defaultXProps.onShow).toHaveBeenCalledTimes(1);
-                expect(result.current.status).toBe(STATUS.OPENING);
-
-                setTimeout(() => {
-                    expect(result.current.status).toBe(STATUS.OPEN);
-
-                    resolve();
-                }, 500);
-            }, 100);
-        });
+        expect(result.current.status).toBe(STATUS.OPEN);
 
         const linkName = 'linkName';
+
         act(() => {
             result.current.handleClose(linkName);
         });
 
-        await new Promise(resolve => {
-            setTimeout(() => {
-                expect(result.current.status).toBe(STATUS.CLOSING);
-
-                setTimeout(() => {
-                    expect(result.current.status).toBe(STATUS.CLOSED);
-
-                    resolve();
-                }, 500);
-            }, 100);
-        });
-
+        expect(defaultXProps.onClose).toHaveBeenCalledTimes(1);
         // onClose handler should be called with the linkName of the method of closing (x button, overlay, etc)
-        expect(defaultXProps.onClose).toHaveBeenCalledWith({ linkName });
+        expect(defaultXProps.onClose).toHaveBeenLastCalledWith({ linkName });
     });
 });